feat: Podcast and Episode details refactor
